model-validator/requirements.txt
requests
PyYAML
pytest-html
pytest-testdox
